+++ app/Http/Controllers/Admin/TaskController.php
@@ -10,34 +10,37 @@ use App\Http\Requests\Admin\Task\TaskIndexRequest;
 use App\Http\Requests\Admin\Task\TaskStoreRequest;
 use App\Http\Requests\Admin\Task\TaskUpdateRequest;
 use App\Models\Company;
-use App\Models\Exhibition;
 use App\Models\Task;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 use Spatie\QueryBuilder\QueryBuilder;
 
 final class TaskController extends Controller
 {
-    public function index(TaskIndexRequest $request, Exhibition $exhibition, Company $company)
+    public function index(TaskIndexRequest $request, Company $company)
     {
+        Gate::authorize('view', $company->exhibition);
+
         $tasks = QueryBuilder::for($company->tasks()->select(['title', 'id', 'deadline', 'status']))
             ->allowedSorts(['title', 'deadline', 'status'])
             ->paginate()
-            ->through(fn ($task): array => [
+            ->through(fn($task): array => [
                 ...$task->toArray(),
                 'status' => $task->status->label(),
             ])
             ->appends($request->query());
 
         return Inertia::render('admin/Tasks/Index', [
-            'exhibition' => $exhibition,
             'company' => $company,
             'tasks' => $tasks,
         ]);
     }
 
-    public function edit(Exhibition $exhibition, Company $company, Task $task)
+    public function edit(Company $company, Task $task)
     {
+        Gate::authorize('view', $company->exhibition);
+
         $user = Auth::user();
         $task->load(['comments' => function ($q) use ($user): void {
             $q->where('user_id', $user->id)
@@ -47,22 +50,24 @@ final class TaskController extends Controller
         }]);
 
         return Inertia::render('admin/Tasks/Edit', [
-            'exhibition' => $exhibition,
             'company' => $company,
             'task' => $task,
         ]);
     }
 
-    public function create(Exhibition $exhibition, Company $company)
+    public function create(Company $company)
     {
+        Gate::authorize('view', $company->exhibition);
+
         return Inertia::render('admin/Tasks/Create', [
-            'exhibition' => $exhibition,
             'company' => $company,
         ]);
     }
 
-    public function store(TaskStoreRequest $request, Exhibition $exhibition, Company $company)
+    public function store(TaskStoreRequest $request,  Company $company)
     {
+        Gate::authorize('view', $company->exhibition);
+
         $task = $company->tasks()->create([
             ...$request->only(['title', 'description', 'deadline']),
             'status' => TaskStatus::TO_BE_COMPLETED,
@@ -82,14 +87,15 @@ final class TaskController extends Controller
             }
         }
 
-        return to_route('admin.exhibitions.tasks.index', [
-            'exhibition' => $exhibition,
+        return to_route('admin.tasks.index', [
             'company' => $company,
         ]);
     }
 
-    public function update(TaskUpdateRequest $request, Exhibition $exhibition, Company $company, Task $task)
+    public function update(TaskUpdateRequest $request, Company $company, Task $task)
     {
+        Gate::authorize('view', $company->exhibition);
+
         $task->update($request->only(['title', 'description', 'deadline']));
 
         if ($request->filled('comment')) {
@@ -119,18 +125,18 @@ final class TaskController extends Controller
             }
         }
 
-        return to_route('admin.exhibitions.tasks.index', [
-            'exhibition' => $exhibition,
+        return to_route('admin.tasks.index', [
             'company' => $company,
         ]);
     }
 
-    public function destroy(Exhibition $exhibition, Company $company, Task $task)
+    public function destroy(Company $company, Task $task)
     {
+        Gate::authorize('view', $company->exhibition);
+
         $task->delete();
 
-        return to_route('admin.exhibitions.tasks.index', [
-            'exhibition' => $exhibition,
+        return to_route('admin.tasks.index', [
             'company' => $company,
         ]);
     }
